Remove now unused strategy
This was the actual cause of the bug in question.
The bug was: when user without cip tries to sign in, they see error message,
but they are actually signed in - they can go to dashboard for example.
The reason for it was 'current_user' method called warden authenticate,
and signed in the resulting user. I think.
The solution was to remove the strategy and only ever sign in users in
sessions controller.