Add docs and config for running jobs locally
Previously there was no active job queue adapter specified so it would use the async adapter:
https://api.rubyonrails.org/classes/ActiveJob/QueueAdapters/AsyncAdapter.html

This meant that the `rake cron:schedule` task would just run the tasks once and not reschedule them
as the async adapter is in-memory.

Set the queue adapter to be delayed_job as it is in other environments and add documentation on
how to run it locally.