Replace PERSONAL_TOKEN in deploy to prod workflow
These steps are failing with 404 errors. Best guess is that this personal token belonged to @kicferk1 who has now been removed from the github org.
We should be able to use GITHUB_TOKEN instead https://docs.github.com/en/actions/security-guides/automatic-token-authentication, and this is
already used in an earlier step in the workflow.